MathServer/requirements.txt
alembic==1.7.4
Flask==2.0.3
Flask-Migrate==3.0.0
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.38.2
psycopg2==2.9.3
python-dotenv==0.20.0
